+++ test/client/get_global_chat_badges_test.dart
@@ -1,0 +1,14 @@
+import 'package:test/test.dart';
+
+import '../utils/twitch_mock_client.dart';
+
+void main() {
+  final client = TwitchMockClient();
+
+  group('getGlobalChatBadges', () {
+    test('1', () async {
+      final badges = await client.getGlobalChatBadges();
+      expect(badges.length, 1);
+    });
+  });
+}